+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ---
 
+### Features
+
+* Array extension
+* Shortcuts section is added to the readme file
+
 ## 0.1.1
 
 ### Features